+++ changelog.md
@@ -3,6 +3,8 @@
 ## [v1.6.2](https://github.com/codeducks/xbot/releases/tag/v1.6.2) (unreleased)
 
 * added config command (old bot command is now info)
+* say hello to chat sanitizer.
+
 
 ## [v1.6.1](https://github.com/codeducks/xbot/releases/tag/v1.6.1) (26/1/2021)